update to match latest dspserver header

+++ N6LYT/ghpsdr3/trunk/src/QtRadio/Spectrum.h
@@ -104,7 +104,7 @@ private:
     int moved;
 
     int sampleRate;
-    int meter;
+    short meter;
     int maxMeter;
     int meterCount;
     int subrx_meter;
